+++ src/test/converter2/issues/gh2234.ts
@@ -1,0 +1,11 @@
+export interface ReadonlyCharMap extends Iterable<string> {
+    at(x: number): string;
+}
+
+export class CharMap implements ReadonlyCharMap {
+    at() {
+        return "";
+    }
+
+    *[Symbol.iterator](): Iterator<string> {}
+}